Requirements for this script to work
For this script to work, download image.png, move it to it's permanent location, edit the script, and update the "BUTTON_IMAGE_PATH " variable to the location of image.png. 

You can take your own screenshot and it'll work just fine. I use the script with a browser that's at 25% zoom and it works just fine.
Press F10 to kill the script.